+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - switched to `kasperl` library for base API and generic pipeline plugins
+- added logging output to the `annotation-overlay-*` and `center-overlay-od`
 
 
 0.0.5 (2025-07-11)